Add a (junk) perform_training function
I wanted to find out how it works to perform updates to a file using the
Colaboratory interface. You can apparently use `<Ctrl-Shift-P>` and
search for `save a copy to github`. Then you authenticate with your
GitHub credentials and choose the correct repository. The file in
question is then updated.